requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistralai==1.10.0
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==2.13.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.52.2
streamlit-drawable-canvas==0.9.3
streamlit-js-eval==0.1.7
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
widgetsnbextension==4.0.15
zipp==3.23.0
google-generativeai==0.8.6
scikit-learn==1.4.0
pytest
pytest-mock